Move definitions out of ensemble so that main isn't called twice
