+++ src/samples/main.ts
@@ -1,7 +1,6 @@
 #!/usr/bin/env node
 
 import {main} from '../lib/app/main.js';
-
-import {sequence1} from './ensemble1.js';
+import {sequence1} from './samples.js';
 
 main(sequence1, []);
